+++ src/frontend/kara.ts
@@ -1,11 +1,8 @@
-import { Vector2 } from './vector.js'
+import { Vector2 } from '../backend/vector.js'
 import { DOM } from './dom_util.js'
 import { draw } from './main.js'
-import {
-	world, mapOverflowToWorld,
-	leaf_positions, shroom_positions,
-	findLeafIndex, findMushroomIndex,
-} from './world.js'
+import { world } from './world.js'
+import { b_world } from '../backend/world.js'
 
 let kara_pos = {
 	x: 0,
@@ -47,12 +44,12 @@ const kara = {
 				x: vec.x + kara_pos.x,
 				y: vec.y + kara_pos.y,
 			}
-			frontField = mapOverflowToWorld(frontField)
+			frontField = b_world.cell_on_world_torus(frontField)
 			let nextField = {
 				x:frontField.x + vec.x,
 				y:frontField.y + vec.y,
 			}
-			nextField = mapOverflowToWorld(nextField)
+			nextField = b_world.cell_on_world_torus(nextField)
 			if (world.isMushroom(nextField.x, nextField.y)) {
 				alert('kara cannot push multiple mushrooms')
 				return
@@ -61,13 +58,12 @@ const kara = {
 				alert('kara tried pushing a mushroom into a tree')
 				return
 			}
-			const idx = findMushroomIndex(frontField.x, frontField.y)
-			shroom_positions[idx] = nextField
+			b_world.relocate_mushroom(frontField.x, frontField.y, nextField)
 		}
 		const vec = getLookVector()
 		kara_pos.x += vec.x
 		kara_pos.y += vec.y
-		kara_pos = mapOverflowToWorld(kara_pos)
+		kara_pos = b_world.cell_on_world_torus(kara_pos)
 		draw()
 	},
 	turnLeft: function(): void{
@@ -101,7 +97,7 @@ const kara = {
 			alert('kara cannot put a leaf on another leaf')
 			return
 		}
-		leaf_positions.push({ x:kara_pos.x, y:kara_pos.y })
+		b_world.push_leaf(kara_pos.x, kara_pos.y)
 		draw()
 	},
 	takeLeaf: function(): void{
@@ -109,12 +105,12 @@ const kara = {
 			karaInactiveAlert()
 			return
 		}
-		const idx = findLeafIndex(kara_pos.x, kara_pos.y)
+		const idx = b_world.find_leaf_index(kara_pos.x, kara_pos.y)
 		if (idx === -1){
 			alert('kara cannot take a leaf where is none')
 			return
 		}
-		leaf_positions.splice(idx, 1)
+		b_world.remove_leaf(kara_pos.x, kara_pos.y)
 		draw()
 	},
 	onLeaf: function(){
@@ -132,7 +128,7 @@ const kara = {
 		let vec = getLookVector()
 		vec.x += kara_pos.x
 		vec.y += kara_pos.y
-		vec = mapOverflowToWorld(vec)
+		vec = b_world.cell_on_world_torus(vec)
 		return world.isTree(vec.x, vec.y)
 
 	},
@@ -159,7 +155,7 @@ const kara = {
 			target.y--
 			break
 		}
-		target = mapOverflowToWorld(target)
+		target = b_world.cell_on_world_torus(target)
 		return world.isTree(target.x, target.y)
 	},
 	treeRight: function(){
@@ -185,7 +181,7 @@ const kara = {
 			target.y++
 			break
 		}
-		target = mapOverflowToWorld(target)
+		target = b_world.cell_on_world_torus(target)
 		return world.isTree(target.x, target.y)
 	},
 	mushroomFront: function(){
@@ -196,7 +192,7 @@ const kara = {
 		let vec = getLookVector()
 		vec.x += kara_pos.x
 		vec.y += kara_pos.y
-		vec = mapOverflowToWorld(vec)
+		vec = b_world.cell_on_world_torus(vec)
 		return world.isMushroom(vec.x, vec.y)
 	},
 	setPosition: function(x: number, y: number) {